requirements.txt
aiogram~=3.21.0
APScheduler~=3.11.0
pytest~=8.4.1
Django~=5.2.4
djangorestframework~=3.16.0